Feature: Turning, Cactus and Background
Added functionality so that the player model switches directions when
left and right movements are inputted. Added a backgorund to the game.
Added a cactus as a danger item. Added collision logic for when player
hits the cactus. Started a life tracking feature - incomplete but when
player hits cactus it loses a life.

To do:
 - Add animations to the player model
 - Add animation for when player model hits cactus (changing velocity
 to backwards direction not working as anticipated)
 - Further develop life system (should lives be handled every frame,
 should it be handled outside the player model)