Adding Animations
I added basic running and jumping animations to the player model. In
doing this I have realized that my code is not structured well, branching
on keys for movement and seperating my code this way may not be optimal.

Furthermore, I still need to look at:
- Death/injury animations
- Seperating the health system to its own script and developing it
further
- using a raybox to detect collisions compared to the singular raycast
being currently used